sftBack/sftMadness/.aws-sam/cache/1ef5c183-e45c-4c93-9cca-9ee15894f336/python/requirements.txt
psycopg2-binary==2.9.10
requests==2.31.0
boto3==1.29.0
botocore>=1.31.44
PyJWT[crypto]==2.8.0
cryptography==36.0.0